CX2: Restore the saved translation
Fill the translation sections with saved translations from earlier
sessions by matching the section number.

This commit implements the very basic section number matching logic.
So it matches a linear order of source sections.

If source article changed heavily, we need to use a better algorithm
to identify source sections even if it changed. And fallback to
original source revision if source article changed beyond restoring
a saved translation. That is for follow up.

Bug: T176843
Change-Id: Idd2dae87b487219d80391ff4f1a9cf74d6e9e8e4