Bring back the overlay support for a specific screen region
Templates and template editor was using this and got
accidentally removed in I0198e67dc2bcf

This commit brings back that feature with the help of an option
fullscreen. The value for this option is by default true, since
that is the general usecase screnario.

Change-Id: Iee396e8517245ec6de2093949c8352015dadb970
Follow-up: I0198e67dc2bcf98e90293d8809aa68b18a220ffb
Bug: T179997