Tweak dump-corpora.php
Instead of doing FormatJson::encode for whole big array, do it
for individual sections instead and stitch it together for one
big file. In other words, home-made streaming json encoder.

This should reduce the memory usage somewhat. If that is not enough
there are further possibilities to improve this script:

1) write to the sink directly, instead of buffering into a string.
2) start streaming to the formatter, so that we do not load all
  sections into the memory before we start formatting them.

Regressions were checked for by comparing the output of this script
before and after this change (default parameters). Only difference
is that this one adds a trailing new line, which is a good thing
in my opinion.

Bug: T180264
Change-Id: Iddbbe85d2db3d444e2ad645bee32fae9bc0ac975