Stop populating ct_params
ContentTranslation used to use the "from" and "to" recorded in ct_params
on its tagged edits for statistics. But this use was removed in
I6f6d026f (February 2015) when CX started recording statistics more
directly in a central database table.

Nothing else anywhere in MediaWiki seems to use the field either. Let's
save a little database space and stop writing it.

Change-Id: Ic514dea72d534ab39c6664f372034641ef882a0e