Logging a warning about section alignment can cause error
While matching cxSourceSection on node.$element, we check if it exists
and if its 'id' property exists. Then, we go on to match the regular
expression. If nothing matches, we log warning. But, while making an
effort to log useful info, we again access same property that might
not exist, according to that code path.

Change-Id: Ie062b054d9353611df8a199aa03fca0f882aad98