Highlight corresponding source section for placeholders
I used the same approach as sentence highlighting. In my testing
MouseOut did not work at all so I removed it, the logic in MouseOver
took care of removing the highlights already. I created more specific
class names for clarity and removed a few things that look unused.

With this approach it is not possible to get stray highlights if one
for example tabs out of the window and moves the cursor away. This
was the case before this patch.

Bug: T187525
Change-Id: I496ada6d14b2791f2112e12584787533ae5c0977