Fix config passing from Iec79e687ef
Fix config passing and clean up after
Iec79e687eff8c63434054282650c97cc882f348d:

* Remove unnecessary 'cx-tools-category-add' message, because VE widget
overrides the message anyway.
* Don't use `allowArbitrary: false` when creating target categories list,
because that parameter is defaulting to false.
* When mw.cx.ui.CategoryMultiselectWidget is created in mw.cx.ui.Categories,
don't pass `this.config` as first parameter to `$.extend`, because that
method is called twice, so config options for source categories list are
persisted when creating config options for target categories, disabling
the reordering feature in the process.
* Config options were meant to be passed from mw.cx.init.Translation all
the way down to mw.cx.ui.CategoryInputWidget, so that siteMapper and
requestManager are reused and not created again. That idea failed, because
when mw.cx.ui.CategoryInputWidget is created in
mw.cx.ui.CategoryMultiselectWidget, only `config.input` is used to extend
the default config option, leaving the desired siteMapper and requestManager
out of the picture.

Change-Id: I990d9aed35a63807aca0a1a043de4caf78238f39