Display dialog to confirm overwrite while publishing
This is meant to improve the current state, where we have JS error
when trying to publish with target title which already exists.
Because of that error which occurs while trying to display the dialog,
user is stuck in publishing with no way out.

Even though this is meant to fix the JS error, there is design to follow,
which is linked in bug.

Title was checked against target wiki, which is now changed to check
against local wiki instead. That is done so consistency is kept with
wiki where page is being published, which uses local wiki as well.

Bug: T190038
Change-Id: I0598006783461c438c8b17d51e78e94263554240