Fix: jQuery.fn.offset() requires an element connected to a document
There is a race condition that can be triggered by changing the target
title from the default value. Instead of trying to tweak the various
change handlers, it seems best for this method to be defensive and
check explicitly that the nodes are attached to the DOM, and not do
anything if they aren't.

Bug: T188077
Change-Id: Ib4c60302e1017e152d9ac3ff1836fb10c82a6036