CX2: Show introduction card only when relevant
As per T188072, the tools column is going to have contextual and
non-contextual tools. Contextual tools are VE tools. Instructions
tool, Article level issues etc are non-contextual tools.

When contextual tools are shown, we won't show non-contextual tools.

This patch takes care of Instructions tools only, but if there are more
non-contextual tools, we need to consider that too.

Bug: T189869
Change-Id: I74c1fb38192fbe0706675b0a279ebca2f50438ff