Fix disabled category tags
When publish button is clicked, many UI items become disabled, and so
does category listing below the content.
OOUI code applies disabled state to the tag multiselect widget and
all tag items. That results in widget and all its tags being disabled
during publishing.
However, when publishing is done, category listing is re-enabled,
and so are all its tags, which breaks intended behavior, which
is to disable unadapted tags in source category listing.

Change-Id: If82f30a46660859c12facc62e01a4e62762e9b1f