Stop shuffling categories
I2e3c6bcbbb8726680296f614944f38f179035578 introduced new behavior for
OO.ui.TagMultiselectWidget, which puts disabled categories at the
beginning.

That is undesired behavior in our case, because we want
to preserve categories order and have it for newly published page.

When publish button is pressed, tag items are disabled, which
makes the parent widget iterate all the items, and disabling them
in a loop. Every time one tag item is disabled, it's moved to the
beginning. That way, we end up with reverse order during and
after publishing.

Change-Id: I89d35472214c486100dabf822180170b1cf512fd