Reduce tools column jump on wide breakpoint
Tools column was limited to 400px in I7b51177418b4. However, that
400px included 20px padding so actual maximum width of tools column
was 380px. That changed to actually be 400px after I474645e72d93.

Breakpoint calculation when distribution between content and tools
columns occurs wasn't pixel perfect because LESS preprocessor cannot
mix em and px values, which are needed to have math for pixel perfect
calculation of breakpoint value. Previous breakpoint was tailored
to previous width value, which now increased.

Since the increase was by 20px, now there's a 20px sudden increase in
size on exact breakpoint, which is:
1200+(0.25*((1200/3)+20)) = 1200+105 = 1305

Previous breakpoint was:
1200+(0.1*((1200/3))) = 1200+40 = 1240

Change-Id: If885b54d8a73cad3170db4cd0b41ddf66b96aeb3