CX2: Fix sticky section highlight
Section highlight works on mouse enter and mouse leave for target
column placeholders. When "Add translation" is clicked and when
placeholder is replaced with MT content, there is no longer a
placeholder element which mouse can leave, so that highlight is
removed from corresonding source section. For that reason, when
paragraph is added to target article, source section remains
highlighted.

Bug: T190908
Change-Id: I2318c708d49a46878a2a45327c8c802c28ac9983