Clean up a bit catch-all config object for main classes
The config object ends up being a poorly documented (if at all) thing
that gets passed around to all constructors. This makes the code very
hard to follow. This patch cleans that up a little by updating the
documentation, making it more explicit what the config object is used
for or what keys are mandatory. Most of the mandatory keys should be
separate parameters, but I only moved one easy case in this patch.

For modified files, also make sure that the license tags and strict
mode are present consistently.

Change-Id: I6816a3190a045df6c45d15dcca43ef44cba50a55