Notify users about unpublished draft deletion
* Introduce new Echo notification type for notifying users that their
old unpublished draft is deleted. If multiple drafts are deleted for
one user, bundle multiple notifications together.
* Clicking on notification can open prefilled "New translation" dialog
to start translating from latest source page revision or open CX
dashboards for user to check in-progress translations.
* If CX dashboard has version param in URL, it will be used to start
translation with that version of Content Translation.
* isExistingTranslation() method in SpecialContentTranslation will
now check if found translation is deleted and return false if so.

Bug: T184268
Change-Id: I1d305d3f97e8cefa96522f77bf9d7c2e523b2af2