Enable publish button only if there's content to publish
* Fix regression from 12f3ba3e7a427c250e9b9b2bb21c900f0375db99 which
enables the publish button when target title doesn't have any errors.
Because of that unconditional enabling, opening new translation gives
same target and source page name, and since source page title is valid,
target title is also valid, which triggers the publish button to be
enabled. For that reason, only enable if there is a content to publish.
* Also, fix old behavior where saving the draft after only changing the
title and then reopening it gives the enabled publish button, because
draft was loaded, but since there is no content, we should not allow
publishing.

Bug: T196463
Change-Id: I6a201a34b5cbe08bea38c7ffe0147d76b4f0c568