Disable all link clicks in source content
Clicking on images in the source should not open them.
That will allow later to select them in order to copy them.
In editor environments (such as Visual Editor), images and
links are selected when clicking them instead of open directly.
Although the source article is not editable in CX, the user
may do other operations on it such as copying content where
opening the content can get in the way.

Control clicks are allowed though.

Bug: T190387
Change-Id: I0f442c676a839e5e60f0b20cf9b4ec6caff04fbe