Revert some parts of fe82d31f3e3bf684c2ccf2cfeeba
Using `this.context.removeItems( [ this ] )` hides link context items.
Other context items, such as images and references are shown properly,
until some piece of text is selected.

This bug was only hidden because ve.ui.CXTextSelectionContextItem was
introduced before fe82d31f3e3bf684c2ccf2cf and was covering the problem.

Use the previous `this.$element.remove()` which was used before this bug
was introduced.

Change-Id: I536bd032ebc45ffdcaec6ae94f9a4b2e9f52a6eb