Correct the check if saved section has validations
Current check is `Object.keys( validations ).length`. Since that code
is executed in a loop, we have section numbers of the sections saved.

Let's say section number is 2.
For that section we can have the following structure of validations:
{
	2: []
}

`Object.keys` returns 1 in this case, so the check passes, and
we execute `onSaveValidation` method, which should not be executed.

Also, current check passes if any other section has some validation,
not need be one we're processing in current loop iteration.

Change-Id: Ic089bfe22be2dbdfaddde3bdcdbc0c970f8370d5