Work around name collisions/duplication of references
When translating a paragraph that contains a reused reference, VE would
crash. This was because the reused reference wouldn't realize it was a
reuse, and would bring along its own copy of the internal item defining
the contents of the reference. VE would then see two references that
both claimed to be the canonical one and pointed to different (but
identical) internal items, but had the same name.

This patch solves this problem in a terribly hacky way, by analyzing the
newFromDocumentInsertion transaction before it's committed and changing
the internal list item pointer for each duplicated reference. We have to
do this after the transaction is generated, because if we try to do it
by modifying the data that goes into newFromDocumentInsertion(), the
internal list index remapping code will clobber the indexes we modified.

Bug: T191011
Change-Id: Ifb2b59e0bb1f8e9f8a16e8088ae3c06083c8902d