Don't miss sections which are edited during saving
mw.cx.TranslationController.prototype.save() adds new section to the
saving queue, which is reset to empty array at the end of
mw.cx.TranslationController.prototype.onSaveComplete()

Because of that, some sections can be falsely treated as saved,
since they are removed from the saving queue and that way we can miss
entire section, or some part of user's edit.

Instead of unconditional setting of saving queue to an empty array,
we get keys of sections which are being saved, and then remove
them upon successful completion of saving process. If some section
was edited during this round of saving, it will be ready for the next go.

Besides that, do some cleanup of mw.cx.TranslationController file:
* Remove usused this.translationId property
* Remove unused 'savestart', 'savesuccess' and 'saveerror' events
* Improve formatting inside this file.

Bug: T199823
Change-Id: Ied274cb16bad2b9f8f181dbd22fddb8fc5ef990a