Translation progress calculation
Introducing mw.cx.TranslationTracker to track the translation
progress at section level and at the whole translation level.

Each section has a state object mw.cx.dm.SectionState. Instances
of that is stored in TranslationTracker indexed with section number.

On section content change, the relative amount of content that
got modified compared to the unmodified MT or source copy. The
calculation is based on number of words that remain unchanged.
Please refer the tests too.

Threshold validation are done and warning are emitted if it
cross the threshold. Sections with less than 10 tokens are
excluded from validations. This is to consider section headings.

Validations are delayed till next change for the placeholder
clicked fresh translations.

The calculated overall progress is now saved too. So that the CX
dashboard displays the translation progress.

When a translation is restored, the section state is restored too.
This is possible because the translation units in our database has
unmodified MT, and MT engine details. Note that the progress at section
level or translation level is recalculated.

Bug: T162113
Bug: T190279
Change-Id: Id767d44b8f80b870ca24ab0436580e94f530454d