Prevent publishing translation with too much unmodified content
Also improves the calculation logic in Translation tracker.

Default threshold set as 99%

Since publish can fail immediately based on this validation, some
UI enable/disable code execution order is changed. Previously,
their order did not matter since publish success or failure
was totally asynchronous.

T190283 need issue cards for this kind of issues, but that is not
included in this ticket. Waiting to finish the issue tracker
refactoring. For now an error message is displayed on top of
translation view.

Bug: T190283
Change-Id: I61ac5d37efe671aff912307aacea381fc2978508