Handle exceptions while notifying users about purged drafts
We store global user IDs in the table, but we are assuming they are
local user IDs when sending notifications. Therefore, use CentralIdLookup
to find local user from central user ID before attempting notification.

Also, log error when object for such user is null, which probably means
that user in anonymous.
If there is an exception while doing notification, catch it,
break from loop for the current user and log some more info.

Bug: T199960
Change-Id: I35fa6fc656e703ac8559f0b5795140a2452c3b33