Fix validation condition
In 9ce1e4cff1b3333f131b8e57d4c528f9c69ead72, condition in
mw.cx.TranslationController.prototype.onSaveComplete, which
controls whether we should go into onSaveValidation was
removed. Instead of that, we deal with empty validation
in onSaveValidation. But, since removal of the condition
meant we let undefined values be passed to onSaveValidation,
we need to check for that as well.

Bug: T202342
Change-Id: Ie3955d94fb45d14be09ee8c9f58be069a6e6be03