Check the text value of translations for progress calculation
This implies that if a section has blank content, it is not counted
as a translated section for progress calculation. So if all sections of
of a translation are blank, the total progress is 0%.

Blank sections will be saved and restored. This is because, even if the
text value is null, the html value is not null. The html value for a
blank section is an empty <section> tag.

Sections become blank either by removing all the content manually or by
choosing the 'Do not use machine translation' option.

Blank sections are now also excluded from unmodified content validation.

Bug: T200996
Change-Id: I7c06ee1f655f915686e10c233b3e95b409df4be0