Fix null being used as source language
If frequent languages, which we get from ULS, are not returning any
valid language code which is different from target language,
define Spanish and English as fallbacks.

Target language is preset to content language, so it's definitely
non-null value.

If target language is English, use Spanish as source language
and if some other language is used as target, fallback to English
as source language.

Spanish is decided as fallback when target language is English
because it's most used as source language for articles published
to English. French is also very close.

Bug: T202286
Change-Id: I4ea8286599db62510c460a229a3a65a39fc5f634