Replace 'redirects: 1' with 'redirects: true'
Although redirects needs truthy value and using 'redirects: 1' is
a shorter form than 'redirects: !0' which we get when minified,
let's use 'redirects: true' because it's used in majority of
other cases.

Also, newbies to the codebase might think 'redirects: 1'
is the number of redirects to get. Or the level of redirects.

Change-Id: I22a6d4ba69b179f37586314484fbfc7596c05c9f