Restore: Use the timestamps of the translation units to choose one
If there is unmodified MT content and user translation, we cannot
blindly choose user translation to restore. There are cases where
user translation is old and the unmodified MT is the latest. This
is due to the selection changes in MT service providers.

This commit uses the timestamp of the records to find which is fresh
content and use that for restoring.

Bug: T203285
Change-Id: I7fd98ffeb0fee535289a41cc468af927d70469fe