Delay the validation of the current paragraph after the user editings
Currently, validations are done for the section immediately after an edit.
Instead of that, Show a warning for the current paragraph, if there
is some time of inactivity after the user started modifying it.

If the user adds a paragraph and starts editing, remains 15 seconds
without making any modification, then the issue is shown.

Bug: T200683
Change-Id: I507e9c35128b7bf1803b453e201acf23fb395287