Section alignment: Do a quick test to determine the alignment need
If the height of source content and target content is same, it is safe
to assume that the sections inside those elements sums up an equal value.
This implies they are aligned in the context of content translation.

I am not seeing a possibility that source column has sections with height
100px, 50px and corresponding sections in target column with 50px, 100px
heights. That would be a case that invalidate the above assumption.

If we go ahead with the same content element height assumption, if the
heights are equal, we don't need to do relatively expensive alignment of
every section in the article.

Currently, for every edit in translation, the whole article alignment
is done in debounced way. But there may be edits that does not change the
height of the parent section.

Again, with the assumption, the cost of alignment is drastically reduced.
Hence reduce the debounce time to 50%.

Bug: T200323
Change-Id: Ibf1fd6208551ca11a87f3930dbff39bb47b42af5