Allow to switch between CX1 and CX2 from the dashboard
* Add a jQuery plugin that adds UI element to perform
the switch between CX1 and CX2
* Redirect the user when translation starts, to include
'&version=1' or '&version=2' in the URL.
Stop relying on front end to decide which version param
should be added to the URL by not using mw.cx.getCXVersion()
when starting new translations.
* When setting the preference, first try using
action: 'globalpreferences' and if that fails
with 'unknown_action' error, try setting the
preference with action: 'options'.

Bug: T194387
Change-Id: Ie4f06c881ba359a80991b04f5dd7fc29cd42facf