For the restored section, give matching id with source section
For the translations where source article changed a lot, we will be
restoring a section with non-matching section number for a source
section. Make sure to set the matching section id for the restored
section.

If the section ids don't match, alignment will go wrong.

Note that we are not talking about parsoid ids. We are talking
about section ids like <section id='cxTargetSection5'>.
We are setting cxTargetSection5 when a section is restored against a
source section <section id='cxSourceSection5'>. It's old Id may be
cxSourceSection4.

If we have a source section like <section id='cxSourceSection5'>
and corresponding target section(restored) as <section id='cxSourceSection4'>,
everything, including seciton alignment will go wrong.
That is what fixed here.

Bug: T200323
Change-Id: I64c05369da32610ab6cff2cc7546d02cd6184963