Add detailed logging for AbuseFilter
For both CX1 and CX2, add a logAbuseFilter() function in the eventlogging
module, which is called after logging the usual publishing error event.

For CX2, move the firing of the mw.cx.translation.publish.error hook
to publishFail(), so that the full JSON data of the error will be available.
That's how that hook was called in CX1, and there's no reason to change
it now. In the future, however, it probably makes sense to gradually retire
the old common event logging for all errors and split it all
to separate schemas.

Bug: T204273
Change-Id: Ie55bf421edc6c3908f5b2874bb4c46f167549d7d