Section alignment: fix height of source section if it is smaller
Since the marginTop property was set, this is not important for
all sections except the last section. For last section, if the
heights are not equal, the category widgets will appear misaligned.

To test, use the article in T203243, after adding to the translation,
do a reset translation since the template re-rendering event is not
triggering section change. Fix for that is coming. After the reset
translation, the alignment should be correct.

Bug: T203243
Change-Id: I64fead51ca9f45f4e0cc6ec18675b1e14f96a337