Wait for autosave if needed while trying to publish
If there are unprocessed changes to the content when user tries to
publish, wait on successful save before actually publishing.
That way, if user decides to make further edits in the future, it
will have the same data as the published translation.
Also, we don't end up with a save after the successful publishing,
which can put the page in 'In-progress' list.

Bug: T200356
Change-Id: I10dda615ce4320da2a88599db36cdecc38b31f95