Introduce getSectionNumber and getSectionId for CXPlaceholderNode
A placeholder node has section id and hence a section number same as
the section that it is going to be replaced with. So, CXPlaceholderNode
can have getSectionNumber and getSectionId methods similar to
CXSectionNode. So, move these methods to CXTranslationUnitModel.

CXLinkAnnotation and CXSentenceSegmentAnnotation classes also inherit
CXTranslationUnitModel. But I don't see any reason to. Probably
a carry forward from previous iterations of CX2. Will fix as follow up.

This fixes T205420 as well. If there is an MT error in a placeholder
click context, getting section number to do prefetch for next section
need not fail without knowing section number. Prevents the JS error
in such contexts.

Bug: T205420
Change-Id: I0b8bb2978c630adcd6d6c813bfcd532a59b7162f