Save: Don't miss unmodified MT while there is modified content
If both unmodified and user translation in save queue.
Only the user translation get saved.

Previously, we were either adding unmodified MT or user modified
content to the records to save. But they are not mutually
exclusive if there was no save happened so far.

Easily reproducible by changing the content on top of an MT
before the first save happens.

Bug: T204101
Change-Id: Ie475a72bb404c0f10c37d714e294dfb0b78f1c7c