Don't convert timestamps for translation units to integer
When translation unit is getting updated, existing row is fetched
from database and object of TranslationUnit class is created by using
that data. In the constructor of that class, timestamp is being
converted from string to integer. On 32-bit environments, timestamp
in format '20181003090915' is overflowing integer limit, which
is 2^31 - 1. While converting the string, it gets capped at that
maximum value. That causes Database::timestamp method to treat
the integer value as Unix timestamp, and timestamp becomes
'20380119031407', which is used as condition to find which
record to update and that clearly always fails.

Bug: T204038
Change-Id: I1b6341085b525c0580571b161252dd13e304a6a8