Give higher rank for ve.dm.MWReferencesListNode than cxTransclusion
The selection of models for a node depends its registration order.

Since ve.dm.MWReferencesListNode also matches for a transclusion node,
the order of registration is used and it fails to get high rank.
This makes reflists considered as transclusions than mwReferencesListNode

In this patch we re-register ve.dm.MWReferencesListNode after the
ve.dm.CXTransclusion* models to allow ve.dm.MWReferencesListNode to
get higher preference.

Bug: T206756
Change-Id: Iea22344b7005de5ea3b1b5f0130388f9aacced9a