Allow adding edit tags to the published articles
Modified the cxpublish api to accept edit tags to add to the
published page. These tags are based on the campaigns configured
for the wiki and campaign parameter in the URL

Note that the campaign param is not stored anywhere for a translation.
So if the translation is restored, this tag won't get applied.

Bug: T208399
Change-Id: I579a8ac83889b3e80c77aba8eef664fe771e11df