Display "publish anyway" dialog if there are translation issues
If the translation has some issues, which are not suppressed,
display the dialog while attempting to publish.
The dialog is meant to get confirmation from users that they
want to publish in spite of existing issues. When translation
with issues is about to overwrite the existing content, the
message is tweaked compared to plain overwrite without issues.

Bug: T190036
Change-Id: I4a88da2ea7713fa26a51d2d57402ddde2f2feef1