Show reset option in the MT card only when section has been modified
Introduce ve.dm.CXSectionNode#hasUserModifications. This requires
access to the translation tracker to know whether the section has
content modifications on top of initial translation.

Known issue 1: for some sections the "Reset translation" option does
not get hidden after used. This is because the translation tracker
thinks there are changes. That is a bug that should be fixed
separately. There seems to be a race condition there where the
isModified flaps between true and false because getUserTranslation
flaps between the previously modified value and fresh value. When
in debugger mode it flaps twice, but last result is correct.

Known issue 2: sections that contain references (or other complex tags)
are always shown as modified if temporarily switched to another engine
and then back. This is because the HTML changes slightly at some point
and no longer matches. Some attributes get reordered and some escaping
of quotes is different.

Bug: T196051
Change-Id: I0c30d4c844d8e5d1c0a5e4eac31ad29500db98c2