Communicate that new translations will be started with CX2
Show the message dialog to inform user that new version of the tool
will be used for new translations that they start.

Should NOT be shown to:
- New users of Content Translation.
- Users who already have manually enabled version 2 of the tool
via CX2 invite or version switcher.

User is considered as new if they haven't started any translations.
To avoid the case where new user starts translation and is no longer
considered new, and then get the message, a condition is added to
take into account only translations started before 2019-01-10.

Bug: T211325
Change-Id: I861eaa2484c30136cb7c5e3b73b74310b4e27616