Reduce the number of JavaScript globals defined on all pages
Moved the ones that are obviously only used by the dashboard or
the translation view to only be defined on those pages.

The remaining ones are used by sitemapper/util classes and it is
more difficult to check whether they are needed on other pages as
well.

See also Id62cf33af45512126f895683fc9ff7c6bc64d08d.

Change-Id: Idfd9d59aa6408165529af909e6c599d2a9137afe