Revert erroneous find-and-replace of page selector focus method call
In 8643c6db0dbbe570e3b8240093780b7caed37b74, eslint config was updated and,
in efforts to adhere to new rules which prohibit usage of jQuery's focus
method directly, find and replace method was used to replace the usages.

Call to this.pageSelector.focus() was replaced by
this.pageSelector.trigger( 'focus' ), which causes
"this.pageSelector.trigger is not a function" error.

This commit reverts the erroneous find-and-replace.

Change-Id: I80af74a7b336370cdab4de72fca13e936a586457