Attach focus listeners only once
36d004f8804daed6ff9575ec9a0540f460c86a09 changed the way we attach
'focus' and 'blur' event handlers for lintable nodes, but previous
logic used to remove all 'mousedown' (was used instead of 'focus')
and 'blur' handlers and then attach the new ones.
With that commit, the logic has changed and such removal of
listeners is gone, which leads to many handlers being attached.
To solve this problem and avoid constant removal and reattachment
of event handlers, introduce a control variable, which lets us
know if handler is already attached, so we don't end up with
many of the same handlers and many cycles of detach-reattach.

Change-Id: Ibd10cdec21d8f8f99ccc4f3613a838837d2bf8ad