Show the invite to add the first paragraph initially
When the translation is empty (and the user is not hovering
any other paragraph in the translation), the invite to add a
new translation will be shown for the first paragraph in the
translation to communicate the starting point to the user.

Some considerations:
- Hovering another paragraph will show the invite only
for the highlighted paragraph.
- Once the user added content to the translation,
no paragraph would be highlighted without the user hovering on it.

Bug: T189203
Change-Id: Idb3fe023e2b8e3c1c183dafd435fe84063e57899