Catch "User account is not global" error
CentralIdLookup method called in Translator::getGlobalUserId(),
returns user ID or 0 if the local user does not correspond to a
central user.
Translator::getGlobalUserId() checks if 0 is returned and throws
an exception. ApiQueryTranslatorStats, which uses this method,
has a condition to check for 0, which is never met, since the
exception is never caught.

Replacing the condition with try-catch block gives us the way
to gracefully return the API error.

Bug: T218902
Change-Id: I3178912af1b033f52da3a652c8eaa6ee87164482