Log need-review events
Grafana is not enough for logging and analyzing
"translation needs review" events, so they should also
be logged to EventLogging for robust analysis.

This reuses the existing ContentTranslation schema,
because it's already deployed. However, it needs a new
type of action, so the schema definition was updated on Meta
and the revision number is updated accordingly.

Bug: T209868
Change-Id: I521e028941f323620c6e0dbbea0f9651750d6449