mw.cx.ui.Header: Avoid wgMainPageTitle for simple main page link
This is the last remaining use of the 'wgMainPageTitle' variable
in Wikimedia Gerrit repositories. This is defined in MediaWiki's
startup module and loaded for all page loads on all wikis.

By using a simple index.php link instead, we don't need to know
the main page title. Instead, MediaWiki figures it out for us.

Bug: T220926
Bug: T219342
Change-Id: Ifdedccb513f5dd4636a37badc329ac5f1c4dd943