Introduce less strict threshold for suppressed MT issues
For paragraphs where the unmodified content warning was shown,
but the user marked it as resolved, apply a less strict threshold
(95% of Machine translation and 75% of source content).
This will provide a way to accommodate cases where the automatic
translation was exceptionally good, but still avoid potential abuse
of the feature (not following blindly the user confirmation).

Bug: T217653
Change-Id: I8e2202a9f34eabf42d48cc5a0bc0abe214511df5