Don't let ULS go outside viewport
Previous code calculated whether it was better to show ULS menu
before or after the language fitler handle, while the menu size
depends on code in ULS itself. So, if menu needed to be large
accoring to ULS code and would go outside when shown after, it
was shown before instead, which could actually make it go outside
viewport even more.

Improve this by always showing the menu before the handle
and keep shrinking the menu if not.
If large menu goes outside viewport when shown before,
make it medium and then narrow if medium is still too big.

Bug: T221950
Change-Id: I702fd09c1247cd19b949de81688d50083e54cd76