Do not use clipboard mode to get HTML of reference list node
Setting clipboard mode helps to have the reference body resolved
and inlined in each paragraph. But if we do the same for reflist,
there are consequences. One issue is the reflist template wrapping
goes away and mw:Extension/references remains.

To test use the testcase in T220491 and observe the published article
has reference list as {{Références}} instead of <references>

Bug: T220491
Change-Id: I6df4ac247e9a0ad6a599d0da323257e84f7c2f71