Use false as default for ContentTranslationDatabase
This config variable is used for setting up database connection,
where it is passed to `getMainLB` method of LBFactory object,
which uses `false` as default for its only `$domain` param.

Other place where this config is used is for update.php hook,
where execution is stopped if this variable is not equal to false,
which is another reason to have `false` as default.

Change-Id: Idecaefd7669cc322bb3ba4daa02e516654e916e2