Adjust publishing restrictions
Adjust publishing restrictions based on the number of
sections affected and previous translations by the user:

In order to support this, consider two factors:
1) Number of sections for which the unmodified content exceeds
the current thresholds.
2) Number of (recently) deleted translations published by the user
in the main namespace.

With those factors, make following adjustments:

For a regular user:
* With 1 problematic section: Allow publish and do not add to the
tracking category (to reduce false positives as described in T217653).
* With 2-9 problematic sections: Allow publishing, but add the
tracking category.
* With 10 or more problematic sections: Prevent publishing.

For a user with deleted translations:
* With 1-4 problematic sections: Allow publishing, but add the
tracking category.
* With 5 or more sections. Prevent publishing.

Bug: T221359
Change-Id: I72e2ed0069480dee6475b66bfe71ae572a209092