Allow searching for titles in any namespace, if query has it
If search query for source page has a namespace, search it in all
namespaces of source wiki. If not, restrict to main namespace.

The presence of namespace in query is just a check for ':' in it.

A query like "Wikipedia:Sig" should return Wikipedia:Signature.
However, "Signature" will not return that result.

Bug: T198205
Change-Id: Iea14ab9ae18473c0c5d06f171bd5fa89c2c60133