Correct popup position when changing publishing destination
When user is not allowed to publish to main namespace, issue
card is shown with action to change title to user namespace.
When user takes that action, popup is displayed to let them
know namespace is changed - T192144.
That popup should point to publish destination settings, but
since e7d17dbc030b91a2d8e55e0760ef10d80419f3ea, publish button
is converted to tool and two buttons make up toolgroup, which
means popup is not pointing to publish settings button,
but to the middle of those two buttons which make toolgroup.

Change-Id: I25be139dab180a580141a7e44d8083848403c090