Fix starting translation through gray interlanguage link on simplewiki
wgSourceLanguage for simplewiki is 'en', and we can extract
'simple' from wiki server name. There was difference in usage of these
two methods, so when gray interlanguage link was hovered on simplewiki,
it correctly showed `from=simple&to=sr`, but clicking to start translation
actually uses source language and we end up with `from=en&to=sr`.
To tackle this problem, create new method that tries getting language
code from domain name and fallback to source language. Instead of
hardcoding the solution for simplewiki, we define hacky mapping for
simplewiki in ContentTranslationDomainCodeMapping config and change
mw.cx.SiteMapper.prototype.getLanguageCodeForWikiDomain to accept
fallback language codes.

Bug: T219470
Change-Id: Icd69bf886f4e92059d28d176e8149f47431202dc