Disallow opening the translation view in the wrong wiki
Because we can only publish articles to the wiki where user has the
translation view open, enforce that the target language matches the
language of current wiki.

Determining the CX language code for current wiki is not trivial,
since it can differ from the content language as well as from the
wiki subdomain. The logic already existed in the purge scripts,
but now I extracted it out to SiteMapper so that it can be reused.

The old getLanguageCode method was only used once, so I replaced it
with the new method since there is no point using those methods
separately.

The validity of the token (special cookie) is not enforced in single
wiki development mode (when $wgContentTranslationTranslateInTarget
is false).

Renamed hasToken to hasValidToken to reflect the increased scope.
Since it had no external callers, I made it private to make it
easier to change it in future.

There is a risk that this will prevent translating to some languages
if our code map is missing entries or they are incorrect. Those can
quickly be fixed by just updating the code map.

Bug: T154888
Change-Id: Iec4126197b7fc775d550a557e5fc34f7a6de954a